+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 * Fixed an issue where running the **format** command on feed integrations removed the `defaultvalue` fields.
 * The **feedReputation** parameter can now be hidden in a feed integration.
+* Fixed an issue where running the **unify** command on JS package failed.
 * Added the *--no-update* flag to the **find-dependencies** command.
 * Added the following validations in **validate** command:
    * Validating that a pack does not depend on NonSupported / Deprecated packs.