+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+* Fixed an issue where running the **format** command on feed integrations removed the `defaultvalue` fields.
+* The **feedReputation** parameter can now be hidden in a feed integration.
+* Fixed an issue where running the **unify** command on JS package failed.
+
+# 1.1.6
 * Added the *--description* option to the **init** command.
 * Added the *--contribution* option to the **init** command which converts a contribution zip to proper pack format.
 * Improved **validate** command performance time and outputs.